refactor: extract HomePage widget to file
